不具合修正
ビルドが通らない箇所を修正。
不要なコードを削除。

+++ src/serialize_base.h
@@ -5,7 +5,7 @@
 
 namespace langscore
 {
-    enum class MergeTextMode : size_t
+    enum class MergeTextMode : int
     {
         AcceptSource = 0,   //元のファイルを残す
         AcceptTarget = 1,   //先のファイルを残す
